backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-dotenv==1.0.0
google-generativeai==0.8.3
pinecone==5.4.1
sentence-transformers==3.3.1
python-multipart==0.0.20
pypdf==5.1.0
